break out and test CaseBlock
